koordlet: fix resctrl cachable updates and core sched UTs
